Refactor chat and interface
