+++ src/vecsync/chat/formatter.py
