+++ src/vecsync/cli.py
@@ -3,7 +3,7 @@ from dotenv import load_dotenv
 from termcolor import cprint
 
 from vecsync.chat.clients.openai import OpenAIClient
-from vecsync.chat.ui import ConsoleUI, GradioUI
+from vecsync.chat.interface import ConsoleInterface, GradioInterface
 from vecsync.settings import Settings
 from vecsync.store.file import FileStore
 from vecsync.store.openai import OpenAiVectorStore
@@ -94,7 +94,7 @@ def sync(source: str):
 def chat_assistant(new_conversation: bool):
     """Chat with the assistant."""
     client = OpenAIClient("test", new_conversation=new_conversation)
-    ui = ConsoleUI(client)
+    ui = ConsoleInterface(client)
     print('Type "exit" to quit at any time.')
 
     while True:
@@ -108,7 +108,7 @@ def chat_assistant(new_conversation: bool):
 @click.command("ui")
 def chat_ui():
     client = OpenAIClient("test")
-    ui = GradioUI(client)
+    ui = GradioInterface(client)
     ui.chat_interface()
